credit_approval_system/requirements.txt
Django==4.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
celery==5.2.7
redis==4.3.4
django-celery-beat==2.4.0
django-celery-results==2.4.0
openpyxl==3.0.10
python-dotenv==0.20.0